requirements.txt
celery==3.1.25
Django==1.9.7
django-celery==3.2.1
flower==0.9.2
kombu==3.0.37
lxml==4.2.3
paramiko==2.4.1
pdfkit==0.6.1
Pillow==5.2.0
pymssql==2.1.3
PyMySQL==0.8.0
pywinrm==0.3.0
redis==2.10.6
reportlab==3.5.5
tornado==5.1.1
xlrd==1.1.0
pyodbc
xlwt==1.3.0
